ec-survey-persist/src/main/java/com/ecarinfo/survey/po/AreaGridInfo.java
package com.ecarinfo.survey.po;
import java.io.Serializable;
import java.util.Date;

public class AreaGridInfo implements Serializable {

	private static final long serialVersionUID = -2260388125919493487L;
	private Integer id;
	private Date createTime;
	private Date updateTime = new Date();

    public Integer getId () {
        return id;
    }

    public void setId (Integer id) {
        this.id = id;
    }

    public Date getCreateTime () {
        return createTime;
    }

    public void setCreateTime (Date createTime) {
        this.createTime = createTime;
    }

    public Date getUpdateTime () {
        return updateTime;
    }

    public void setUpdateTime (Date updateTime) {
        this.updateTime = updateTime;
    }
}